+++ Sources/PrintingKit/PrinterView.swift
@@ -11,10 +11,25 @@ import SwiftUI
 /**
  This protocol can be implemented by any view that should be
  able to print ``PrintItem`` types.
+ 
+ Implementing this protocol gives the view access to a bunch
+ of convenient print functions.
  */
 public protocol PrinterView: View {}
 
 public extension PrinterView {
+    
+    /**
+     Whether or not the view can print images.
+     
+     - Parameters:
+       - printer: The printer to use, by default `.standard`.
+     */
+    func canPrintImages(
+        with printer: Printer = .standard
+    ) -> Bool {
+        return printer.canPrintImages
+    }
     
     /**
      Whether or not the view can print a certain print item.
@@ -32,6 +47,18 @@ public extension PrinterView {
     }
     
     /**
+     Whether or not the view can print views.
+     
+     - Parameters:
+       - printer: The printer to use, by default `.standard`.
+     */
+    func canPrintViews(
+        with printer: Printer = .standard
+    ) -> Bool {
+        return printer.canPrintViews
+    }
+    
+    /**
      Print a certain print item.
      
      - Parameters:
@@ -43,5 +70,45 @@ public extension PrinterView {
         with printer: Printer = .standard
     ) throws {
         try printer.print(item)
+    }
+    
+    /**
+     Try to print the provided view as an image.
+     
+     - Parameters:
+       - view: The view to print.
+       - scale: The scale to print in, by default `2`.
+     */
+    @MainActor
+    @available(iOS 16.0, macOS 13.0, *)
+    func printView<Content: View> (
+        _ view: Content,
+        withScale scale: CGFloat = 2,
+        printer: Printer = .standard
+    ) throws {
+        let item = try PrintItem.view(view, withScale: scale)
+        try printer.print(item)
+    }
+    
+    /**
+     Try to print the provided view as an image.
+     
+     - Parameters:
+       - view: The view to print.
+       - scale: The scale to print in, by default `2`.
+     */
+    @available(iOS 16.0, macOS 13.0, *)
+    func printViewAsTask<Content: View> (
+        _ view: Content,
+        withScale scale: CGFloat = 2,
+        with printer: Printer = .standard
+    ) {
+        Task {
+            try? await printView(
+                view,
+                withScale: scale,
+                printer: printer
+            )
+        }
     }
 }