+++ Sources/PrintingKit/StandardPrinter_iOS.swift
@@ -19,6 +19,10 @@ public class StandardPrinter: Printer {
     
     public init() {}
     
+    public var canPrintImages: Bool {
+        return true
+    }
+    
     public func canPrint(_ item: PrintItem) -> Bool {
         switch item {
         case .imageData(let data): return data.canCreateExportFile